Create some basic network discovery and connectivity functionality
The application can look for peers, connect to them, and display their connection info in the logs. It has a background service that performs this work and can shut down cleanly when necessary. A connection manager class is used to maintain a list of devices we're already connected to so that connection attempts aren't made twice to the same device.

What is still left to be done is actual network transmission to these devices, as well as handling failure cases such as a random disconnection due to signal loss.