Fix some ConcurrentModificationExceptions by just making accesses synchronized
It's not going to take a long time to read the list of connections - it's not loading anything from disk or network, just looking through a local list. As such, it's probably fine to just have synchronized access to it to prevent exceptions while accessing from multiple threads. Creating a concurrent solution would actually be a lot harder.