Start conversion to Websockets
Major changes to the project are going to be made, mainly the use of Websockets and a Websocket library instead of a custom TCP server/client.

Also, the binary transport protocol may disappear in a future version and be replaced with something better, like JSON/BSON payloads with data in them; many of the packets in the protocol were meant to do something that the Websocket library can handle.